change reading validity to status flags; merge logic updated, only merge gps/cpm if its available; updated checks before sending to api, reading needs to be valid ¨

+++ bgeigiecast/dms_dd.cpp
@@ -1,14 +1,14 @@
 #include "dms_dd.h"
 
 
-double dms_to_dd(float dms) {
+double dms_to_dd(double dms) {
   double degree = static_cast<int>(dms / 100);
   double minutes = static_cast<int>(dms - (degree * 100));
   double seconds = static_cast<int>((dms - (degree * 100) - minutes) * 100);
   return degree + minutes/60 + seconds/3600;
 }
 
-double dm_to_dd(float dm) {
+double dm_to_dd(double dm) {
   double degree = static_cast<int>(dm / 100);
   double minutes = static_cast<int>(dm - (degree * 100));
   return degree + minutes/60;
